refactor(types): move types from root type def file to just the vec module/directory
